refactor: re-home GitHub Actions TestOps adapter to providers/github (CHAOS-1556)
